// src/apis/entityConfig.tsx
export interface EntityConfigDto {
  id?: string;
  friendlyName?: string | null;
  typeShortAlias?: string | null;
  tableName?: string | null;
  className?: string | null;
  namespace?: string | null;
  discriminatorValue?: string | null;
}